requirements.txt
torch
transformers
pillow
faiss-cpu
streamlit
numpy
